test: consolidate per-resource test files; drop Phase11 naming
Folds MoviesPhase11Tests / MusicPhase11Tests / GamesPhase11Tests into
the corresponding *EndpointsTests files. Each resource now has one
test class organised by what it tests (auth, CRUD, ownership,
validation, list filters, fields round-trip, tags) via section
comments rather than by when the tests were written.

Each file gets one comprehensive Response record covering every field
the endpoint returns, and one Sample() factory with sensible defaults
that all tests share. Coverage and test count are unchanged (84/84).

The phase number was implementation history; future readers should not
have to know which release added a particular field to find its test.
